+++ src/componentes/formsInputs/registerInputs.js
@@ -1,37 +1,86 @@
-import React from 'react';
+import React from "react";
 import TextField from "@mui/material/TextField";
 
 function RegisterInputs({
-  setDados, type, label, id, field, select, required, initialErrorState
-
+  setDados,
+  type,
+  label,
+  id,
+  field,
+  select,
+  required,
+  initialErrorState,
+  mask,
+  dados,
 }) {
-
-  const handleChange = (e, entrada) => {
-    const { value } = e.target;
+  const handleChange = (value, entrada) => {
     setDados(value, entrada);
   };
 
   return (
-    <TextField 
-    required={required}
-    id={id}
-    error={initialErrorState[`${id}`]}
-    onChange={(e) => handleChange(e, id)}
-    label={label}
-    type={type}
-    select={select}
-    SelectProps={{
-        native: true,
-      }}
-    variant="standard"
-    >
-        {field
-        && field.map((option) => (
-            <option key={option.value} value={option.value}>
+    <div>
+      {type === "date" && (
+        <TextField
+          required={required}
+          id={id}
+          error={initialErrorState[`${id}`]}
+          value={dados[`${id}`]}
+          onChange={(e) => handleChange(e.target.value, id)}
+          label={label}
+          InputLabelProps={{shrink:true}}
+          type={type}
+          variant="standard"
+        />
+      )}
+      {mask && (
+        <TextField
+          required={required}
+          id={id}
+          error={initialErrorState[`${id}`]}
+          value={dados[`${id}`]}
+          onChange={(e) => handleChange(mask(e.target.value), id)}
+          label={label}
+          type={type}
+          select={select}
+          variant="standard"
+        />
+      )}
+      {select && (
+        <TextField
+          required={required}
+          id={id}
+          error={initialErrorState[`${id}`]}
+          value={dados[`${id}`]}
+          onChange={(e) => handleChange(e.target.value, id)}
+          label={label}
+          SelectProps={{
+            native: true,
+          }}
+          type={type}
+          select={select}
+          variant="standard"
+        >
+          {field &&
+            field.map((option) => (
+              <option key={option.value} value={option.value}>
                 {option.label}
-          </option>
-        ))}
-    </TextField>
+              </option>
+            ))}
+        </TextField>
+      )}
+      {!mask && !(type === "date") && !select && (
+        <TextField
+          required={required}
+          id={id}
+          error={initialErrorState[`${id}`]}
+          value={dados[`${id}`]}
+          onChange={(e) => handleChange(e.target.value, id)}
+          label={label}
+          type={type}
+          variant="standard"
+        />
+      )}
+    </div>
   );
 }
 export default RegisterInputs;
